\name{BP}
\alias{BP}
\docType{data}
\title{BaySeq Profile used in simulations by Hardcastle et al}
\description{
  RNA-seq profile datasets.
}
\usage{data(BP)}
\format{
  Data frames with 8647 observations on the following 2 variables.
  \describe{
    \item{\code{gene}}{a character vector}
    \item{\code{expression}}{a numeric vector}
  }
}
\references{
 Hardcastle, T.J. & Kelly, K.a. baySeq: empirical Bayesian methods for identifying differential expression in sequence count data. BMC bioinformatics 11, 422-422.
}
\keyword{datasets}