\name{HBRmodel}
\alias{HBRmodel}
\docType{data}
\title{Average abundance for RNA-seq data from Human Brain Reference.}
\description{
  RNA-seq datasets.
}
\usage{data(HBRmodel)}
\format{
  Data frames with 17597 observations on the following 2 variables.
  \describe{
    \item{\code{GeneName}}{a character vector}
    \item{\code{Count}}{a numeric vector}
  }
}
\references{
  Au, K.F., Jiang, H., Lin, L., Xing, Y. & Wong, W.H. Detection of splice junctions from paired end RNA-seq data by SpliceMap. Nucleic Acids Res 38, 4570-4578.
}
\keyword{datasets}
