\name{SingleCell}
\alias{SingleCell}
\docType{data}
\title{Single-cell RNA-seq data for model free simulation}
\description{
  Single-cell RNA-seq.
}
\usage{data(SingleCell)}
\format{
  Data frames with 51516 rows and 96 columns.
  \describe{
    \item{Details}{This is single-cell RNA-seq data from which counts were generated when using RNA-seq model free approach.}
    
  }
}
\references{
  In-house data.
}
\keyword{datasets}
