\name{Wu}
\alias{Wu}
\docType{data}
\title{Average abundance for RNA-seq data from schizophrenia.}
\description{
  RNA-seq profile datasets.
}
\usage{data(Wu)}
\format{
  Data frames with 18982 observations on the following 2 variables.
  \describe{
    \item{\code{gene}}{a character vector}
    \item{\code{expression}}{a numeric vector}
  }
}
\references{
  Wu, J.Q. et al. Transcriptome sequencing revealed significant alteration of cortical promoter usage and splicing in schizophrenia. PloS One 7, e36351-e36351.
}
\keyword{datasets}